`near-network`: Move types to `network_protocol.rs`
