+++ chain/network-primitives/README.md
@@ -1,0 +1,9 @@
+# Overview
+TODO - already in another PR
+
+# Project structure
+- `network_protocol.rs` - contains types, which are part of network protocol, they should be changes with care. 
+All of them derive `BorhSerialize` / `BorshDeserialize`
+- `types.rs` - TODO
+- `config.rs` - TODO
+- `actix.rs` - TODO
