+++ chain/network-primitives/src/lib.rs
@@ -1,2 +1,3 @@
+mod network_protocol;
 pub mod types;
 pub mod utils;
